+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/content/media/TextedMediaGroupMediaContent.kt
@@ -1,0 +1,3 @@
+package dev.inmo.tgbotapi.types.message.content.media
+
+sealed interface TextedMediaGroupMediaContent : TextedMediaContent, MediaGroupContent