+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/content/media/TextedMediaGroupMediaInput.kt
@@ -1,3 +1,0 @@
-package dev.inmo.tgbotapi.types.message.content.media
-
-sealed interface TextedMediaGroupMediaInput : TextedMediaContent, MediaGroupContent