+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/content/media/VisualMediaGroupContent.kt
@@ -2,6 +2,6 @@ package dev.inmo.tgbotapi.types.message.content.media
 
 import dev.inmo.tgbotapi.types.InputMedia.VisualMediaGroupMemberInputMedia
 
-interface VisualMediaGroupContent : TextedMediaGroupMediaInput {
+interface VisualMediaGroupContent : TextedMediaGroupMediaContent {
     override fun toMediaGroupMemberInputMedia(): VisualMediaGroupMemberInputMedia
 }